ai-server/requirements.txt
tensorflow==2.19.0
tqdm
pillow
flask